statusdata/admin.py
from django.contrib import admin
from . import models
from import_export.admin import ImportExportModelAdmin

#Check existance model starts
@admin.register(models.Enquiry)
class EnquiryAdmin(ImportExportModelAdmin, admin.ModelAdmin):
    list_display = ['prospect_name', 'mobile', 'email', 'model']


@admin.register(models.Booking)
class BookingAdmin(ImportExportModelAdmin, admin.ModelAdmin):
    list_display = ['prospect_name', 'mobile', 'email', 'model']


@admin.register(models.Retail)
class RetailAdmin(ImportExportModelAdmin, admin.ModelAdmin):
    list_display = ['delivery_date','model','cust_id','prospect_name','dse_code']


@admin.register(models.RetailCancel)
class RetailCancelAdmin(ImportExportModelAdmin, admin.ModelAdmin):
    list_display = ['delivery_date','prom_delivery_date','model','cust_id','prospect_name','dse_code']


@admin.register(models.Delivery)
class DeliveryAdmin(ImportExportModelAdmin, admin.ModelAdmin):
    list_display = ['prospect_name', 'mobile', 'email', 'model']

#Check existance model ENDS