requirements.txt
Django==4.1.2
djangorestframework
django-mysql==2.2.0
setuptools
channels
tagalog
django-admin-volt
crispy-bootstrap5
django-filter
drf-nested-routers
django-debug-toolbar
djoser
pillow
django-filter
django-import-export
django-cors-headers
